new_framework/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
cryptography>=41.0.0
APScheduler>=3.10.0
requests>=2.31.0
aiofiles>=23.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
